Refactor App component to use custom hooks for movie fetching and local storage management; implement useKey hook for keyboard event handling
