+++ src/useLocalStorageState.js
@@ -1,0 +1,15 @@
+import { useEffect,useState } from "react";
+
+export function useLocalStorageState(initialItems,key){
+    const [value, setValue] = useState(function () {
+    const storedValue = localStorage.getItem(key);
+    return storedValue ? JSON.parse(storedValue) : initialItems;
+  });
+  useEffect(
+    function () {
+      localStorage.setItem(key, JSON.stringify(value));
+    },
+    [value,key]
+  );
+  return[value,setValue]
+}
